cleanup: remove support for gdb 5.x
The last release of gdb 5.x was in 2002. We can remove support for it,
cleaning up the code. Since setup_gdb_defaults is using its own strings,
we can leave the block for this version empty to avoid messing with the
version #defines that are used as indexes. We'll clean it up in a later
patch.
(jeffm@suse.com)