cleanup: use minimal_symbol directly instead of pointers to longs
patch_kernel_symbol always deals with minimal symbols. Just pass the pointer
in and back out to a callback that uses the right macros. This isn't
entirely necessary for use with gdb 7.6 but will be for gdb 7.8 since
it doesn't provide a single value that can be used to generate a pointer.
(jeffm@suse.com)