cleanup: use the gdb interpreter infrastructure
The hook assignment method of providing an alternative interface has been
deprecated from gdb for some time. Gdb 7.7 or 7.8 removed the
deprecated_command_loop_hook callback. It's time to just use the
interpreter interface. Since all we really want to do is provide a
different event loop, we can copy the cli interpreter and override the
command_loop_proc there. In doing so, we can also let gdb initialize
it for us, get rid of the separate newline-and-exit callback for
version queries, and drop the update_gdb_hooks callback.
(jeffm@suse.com)