Fix for the handling of compound pages in Linux 4.4 and later kernels, which contain this kernel commit:
  commit 1d798ca3f16437c71ff63e36597ff07f9c12e4d6
  mm: make compound_head() robust

The commit above removes the PG_tail and PG_compound page.flags bits
and the page.first_page member, and introduces a page.compound_head
member, which is a pointer to the head page and whose bit 0 acts as
the tail flag.  Without the patch, a SLAB or SLUB warning message
that indicates "cannot determine how compound pages are linked" is
displayed during initialization, and any command that tracks compound
pages will be affected.
(anderson@redhat.com)