Allows to change the error output direction
Currently, the error() is always printing the output to the console
through 'stdout'. This does not follow redirection which is good when
you want to know error while redirecting commands output to a file.
However, there are situations that you want to hide error messages or
redirect it into somewhere else.

Using 'set stderr' command, it can be changed to three different
destination - fixed 'stdout', following redirection (fp), or a custom
file path.

    crash> set stderr
    stderr: stdout
    crash> sym 0x523 > /dev/null
    sym: invalid address: 0x523
    crash> set stderr fp
    stderr: fp
    crash> sym 0x523 > /dev/null
    crash> set stderr /tmp/err.log
    stderr: /tmp/err.log
    crash> sym 0x523 > /dev/null
    crash> set stderr stdout
    stderr: stdout
    crash> sym 0x523 > /dev/null
    sym: invalid address: 0x523