extensions/trace: Sync struct member names with upstream kernel
Trying to use the trace extension on a mainline kernel (v5.8-rc4) doesn't
work, and it seems to stem from some struct renaming that has happened
upstream.

ftrace's ring_buffer got renamed to trace_buffer in v5.6 by commit

  13292494379f ("tracing: Make struct ring_buffer less ambiguous")

and struct trace_array's trace_buffer got renamed to array_buffer also in
v5.6 by commit

  1c5eb4481e01 ("tracing: Rename trace_buffer to array_buffer")

Sync the trace extension with those changes.

Note that this isn't backwards-compatible with pre-v5.6 kernels, I don't
know if crash is expected to be aligned with the latest kernels or if we
need to throw in a bucketload of #ifdefs to make this work somehow.

Signed-off-by: Valentin Schneider <valentin.schneider@arm.com>