scope: add ability to un-set scope
Currently there is no way to un-set the scope without having to
exit and re-enter crash.  The ability to un-set can come in very
useful when running automated pykdump scripts and needing scope to
be cleared between script runs.  Add the ability by allowing
'set scope 0' as a viable command.

Signed-off-by: John Pittman <jpittman@redhat.com>