Fix "sys [-t]|mod -S" after "mod -t" when crash runs with -s option
When crash runs with -s option, SIZE(taint_flag) and OFFSET(tnt_false)
are not set during initialization.  If the "mod -t" option is executed,
it sets the former but does not set the latter.  After that, the "sys"
command uses OFFSET(tnt_false) without setting it, because it checks
only whether SIZE(taint_flag) is set.

Without the patch, the "sys [-t]" and "mod -S" options after "mod -t"
option fail with the error message:

  sys: invalid structure member offset: tnt_false
       FILE: kernel.c  LINE: 11203  FUNCTION: show_kernel_taints_v4_10()

Signed-off-by: Kazuhito Hagio <k-hagio-ab@nec.com>