Makefile: reduce crash build log
Currently the verbose output of tar command when extracting the
GDB source files occupies more than the half of crash build log.
It is not so helpful and makes the build log longer needlessly
especially on CI build test.  Let's stop it.

Also reduce about 600 lines of crash build log with wget's
"--progress=dot:mega" option when stderr is not a tty.

Signed-off-by: Kazuhito Hagio <k-hagio-ab@nec.com>