struct: fix struct print member array of list_heads
Due to the way that an array of list_head entries are printed,
parsing of them fails.  Note the difference in spacing between the
double opening and double closing brackets.

    crash> struct blk_mq_ctx.rq_lists ffffc447ffc0f740
    <-->rq_lists = {{
            next = 0xffffc447ffc0f748,
            prev = 0xffffc447ffc0f748
          }, {
            next = 0xffffc447ffc0f758,
            prev = 0xffffc447ffc0f758
          }, {
            next = 0xffffc447ffc0f768,
            prev = 0xffffc447ffc0f768
    <---->}}

As parse_for_member() relies on opening and closing brackets having
the same spacing, make a condition for these arrays of list_head
members.

    Before:

    crash> struct blk_mq_ctx.rq_completed ffffc447ffc0f740
    crash>

    After:

    crash> struct blk_mq_ctx.rq_completed ffffc447ffc0f740
      rq_completed = {221, 1333}

Signed-off-by: John Pittman <jpittman@redhat.com>