MIPS64: Add 'help -r' command support
Add support form printing out the registers from the dump file. We don't
take the registers directly from the ELF notes but instead use the version
we've saved into the machine_specific structure.  If we don't do this,
we'd get misleading output when the number of ELF notes don't match the
number of online CPUs.

E.g. Without this patch:
crash> help -r
help: -r option not supported for this dumpfile

E.g. With this patch:
crash> help -r
...
CPU 3:
     R0: 0000000000000000   R1: 0000000000000001   R2: 9800000254f3c400
     R3: 0000000000000000   R4: ffffffff8123b6b0   R5: 0000000000000000
     R6: 9800000243bcf200   R7: fffffffffffffff8   R8: fffffffffffffffd
     R9: 00180000000000ff  R10: ffffffff810e29a8  R11: ffffffff80e84190
    R12: 000000005400cce0  R13: 0000000000000000  R14: 0000000000000040
    R15: 6e69636e79732074  R16: ffffffff81200000  R17: ffffffff81240000
    R18: 0000000000000000  R19: ffffffff81430000  R20: 980000024291f938
    R21: 0000000000000001  R22: 980000024aef7320  R23: ffffffff81430000
    R24: 0000000000000002  R25: ffffffff80878b58  R26: 0000000000000000
    R27: 0000000000000000  R28: 980000024291c000  R29: 980000024291f930
    R30: 9800000243bcf200  R31: ffffffff802fff00
       LO: ffffffff81210000        HI: ffffffff81210000
      EPC: ffffffff802fff84  BADVADDR: ffffffff802bbe9c
   STATUS: ffffffff81430000     CAUSE: 9800000243bcf200

Signed-off-by: Huacai Chen <chenhuacai@loongson.cn>
Signed-off-by: Youling Tang <tangyouling@loongson.cn>