GDB: fix the failure of 'set scope' command
Currently some commands such as 'sys' may cause subsequent 'set scope'
commands to fail because it may not find the correct symtab associated
with PC and SECTION in the find_pc_sect_symtab(), eventually, this will
cause the following failure:

  crash> mod -S 3.10.0-957.el7.x86_64
  crash> mod -d dm_service_time
  crash> mod -sr dm_service_time
  crash> set scope st_create
  scope: ffffffffc044d270 (st_create)
  crash> sys
        KERNEL: 3.10.0-957.el7.x86_64/vmlinux
      DUMPFILE: crash/vmcore  [PARTIAL DUMP]
  ...
  crash> set scope st_create
  set: gdb cannot find text block for address: st_create

To find the correct symtab, let's check whether there is an address
mapping to 'block' in the symtab searching loop and the PC is in the
range.  If the symtab associated with PC is found, and then use it.

Signed-off-by: Lianbo Jiang <lijiang@redhat.com>