symbols: fix offset print for function pointers that return pointers
In the show_member_offset() function, when trying to handle function
pointers, the case for "(*" is handled.  However, if the function
pointer returns a pointer or a pointer to a pointer, then the
condition is unhandled.  This results in the offset not being printed,
for example:

  crash> struct -o offload_callbacks
  struct offload_callbacks {
         struct sk_buff *(*gso_segment)(struct sk_buff *, netdev_features_t);
         struct sk_buff **(*gro_receive)(struct sk_buff **, struct sk_buff *);
    [16] int (*gro_complete)(struct sk_buff *, int);
  }

Fix by first checking if the member is potentially a function pointer,
then checking if it returns a pointer or a pointer to a pointer.

[ kh: added the example output above ]

Signed-off-by: John Pittman <jpittman@redhat.com>