Skip HZ calculation using cfq_slice_async in Linux 4.8+
In task_init() there is a calculation of machdep->hz based on the
value of cfq_slice_async if the symbol exists.

However, kernel commit 9a7f38c42c2b ("cfq-iosched: Convert from
jiffies to nanoseconds") changed the definition of cfq_slice_async
from (HZ / 25) to (NSEC_PER_SEC / 25).  As such, the HZ calculation
will result in a value of 1000000000 for machdep->hz.   In vmcores
where the symbol exists and has this definition, this causes incorrect
results for some calculations.  For a couple of 4.12 vmcores in this
situation crash shows the uptime as 3 seconds, which also throws off
the timestamps in "log -T".

Fix this by skipping the code block for Linux 4.8 and above.

Signed-off-by: Martin Moore <martin.moore@hpe.com>