memory: Fix for "kmem -n" option to display NID correctly
The nid member of struct memory_block is a 4-byte integer, but read
and printed as a 8-byte integer on 64-bit machines.  Without the
patch, the option displays wrong NIDs.

  crash> kmem -n
  ...
     MEM_BLOCK        NAME          PHYSICAL RANGE       NODE  STATE   START_SECTION_NO
   ffff9edeff2b9400   memory0             0 -   7fffffff 14195095130662240256  ONLINE  0
   ffff9edeff2bb400   memory2     100000000 -  17fffffff 14195094718345379840  ONLINE  32

The issue seems to appear on Linux 5.12 and later kernels that contain
commit e9a2e48e8704c ("drivers/base/memory: don't store phys_device
in memory blocks"), which changed the arrangement of the members of
struct memory_block.

Signed-off-by: Kazuhito Hagio <k-hagio-ab@nec.com>