Fix the failure of reporting vmcore and vmlinux do not match for kernels(<2.6.11)
There is a regression issue for kernels(<2.6.11) as below:

    $ crash 2.6.9-68.9/vmcore 2.6.9-68.9/vmlinux.gz
    ...
    GNU gdb (GDB) 10.2
    ...
    crash: /var/tmp/vmlinux.gz_GLsAvX and 2.6.9-68.9/vmcore do not match!

The reason is that it needs to read out the address of linux banner with
readmem() first, and then the read_string() will be able to read the data
from linux banner. So, for the kernels(<2.6.11) case, lets still invoke
get_symbol_data() to accomplish this. See the changes:
[1] https://elixir.bootlin.com/linux/v2.6.10/source/init/version.c#L38
[2] https://elixir.bootlin.com/linux/v2.6.11/source/init/version.c#L38

Signed-off-by: Tao Liu <ltao@redhat.com>
Signed-off-by: Lianbo Jiang <lijiang@redhat.com>