log: add warning to help text to inform the inaccuracy of -T option
The timestamps of the "log -T" option are inaccurate because they are
from local_clock(), which returns the raw counter in the local CPU and
it's different from the elapsed wall time.

The dmesg command, which the "log -T" option imitates, has a similar
behavior in nature and a warning in its help text.  Let's add a warning
also to the crash's help text to inform the inaccuracy for now.

Link: https://listman.redhat.com/archives/crash-utility/2021-September/msg00044.html
Reported-by: Martin Moore <martin.moore@hpe.com>
Signed-off-by: Kazuhito Hagio <k-hagio-ab@nec.com>