x86_64_init: Refresh vmalloc region addresses in POST_RELOC instead of POST_GDB phase
Previously for x86_64, when memory is randomized, the region addresses
such as vmalloc_start_addr/vmemmap_vaddr/modules_vaddr are firstly set
to a default value before POST_RELOC phase, then get refreshed with the
actual value in POST_GDB phase.

However for crash mininal mode, POST_GDB phase is not called, which
leaving the region addresses unrefreshed and incorrect. As a consequence,
the x86_64_IS_VMALLOC_ADDR check will give a faulty result when
value_search tries to search a symbol by address.

For example, in crash minimal mode we can observe the following issue:

    crash> dis -f panic
    dis: cannot resolve address: ffffffffb20e0d30

    crash> sym panic
    ffffffffb20e0d30 (T) panic /usr/src/debug/kernel-4.18.0-290/linux-4.18.0-290/kernel/panic.c: 168
    crash> sym ffffffffb20e0d30
    symbol not found: ffffffffb20e0d30

In this patch, we will move the code which update the region addresses into
POST_RELOC phase, so in mininal mode the regions can get the correct
addresses.

Signed-off-by: Tao Liu <ltao@redhat.com>