Fix the failure of resolving ".rodata" on s390x
The commit <cd8954023bd4> broke crash-utility on s390x and got the
following error:

  crash: cannot resolve ".rodata"

The reason is that all symbols containing a "." may be filtered out
on s390x. To prevent the current failure, do not filter out the
symbol ".rodata" on s390x.

In addition, a simple way is to check whether the symbol ".rodata"
exists before calculating the value of a symbol, just to be on the
safe side.

Fixes: cd8954023bd4 ("kernel: fix start-up time degradation caused by strings command")
Reported-by: Alexander Egorenkov <egorenar@linux.ibm.com>
Signed-off-by: Lianbo Jiang <lijiang@redhat.com>