sbitmapq: Limit kernels without sbitmap again
commit 364b2e413c69 ("sbitmapq: remove struct and member validation
in sbitmapq_init()") allowed the use of the "sbitmapq" command
unconditionally.  Without the patch, the command fails with the
following error on kernels without sbitmap:

  crash> sbitmapq ffff88015796e550

  sbitmapq: invalid structure member offset: sbitmap_queue_sb
            FILE: sbitmap.c  LINE: 385  FUNCTION: sbitmap_queue_context_load()

Now the command supports Linux 4.9 and later kernels since it was
abstracted out, so it can be limited by the non-existence of the
sbitmap structure.

Signed-off-by: Kazuhito Hagio <k-hagio-ab@nec.com>