Fix "task -R" by adding end identifier for union in task_struct
Previously, the start and end identifiers for union are "  {\n" and
"  }, \n".  However the end identifier is not always as expected.
"  },\n" can also be the end identifier with gdb-10.2.  As a result,
variable "randomized" is in incorrect state after union, and fails to
identify the later struct members.  For example, we can reproduce the
issue as follows:

    crash> task
    PID: 847      TASK: ffff94f8038f4000  CPU: 72   COMMAND: "khungtaskd"
    struct task_struct {
      thread_info = {
	flags = 2148024320,
	status = 0,
	preempt_lazy_count = 0
      },
      {
	<the union>
      },
      ...
      wake_entry = {
	next = 0x0
      },
      ...

Before patch:

    crash> task -R wake_entry
    PID: 847      TASK: ffff94f8038f4000  CPU: 72   COMMAND: "khungtaskd"

After patch:

    crash> task -R wake_entry
    PID: 847      TASK: ffff94f8038f4000  CPU: 72   COMMAND: "khungtaskd"
      wake_entry = {
	next = 0x0
      },

Signed-off-by: Tao Liu <ltao@redhat.com>