Fix segmentation fault in page_flags_init_from_pageflag_names()
When read_string() fails in page_flags_init_from_pageflag_names(),
error() dereferences the name variable to print the string that the
variable points to. However, name points to a string that is not in
crash's memory-space thus triggering a segmentation fault.

This patch replaces "%s" in the error message with "%lx" so the address
is printed instead.  Also replaces "%ld" for mask with "%lx".

[ kh: changed the conversion specifiers and commit message ]

Signed-off-by: Matias Ezequiel Vara Larsen <matias.vara@vates.fr>
Signed-off-by: Kazuhito Hagio <k-hagio-ab@nec.com>