Fix for "ps/vm" commands to display correct %MEM and RSS values
The ps/vm commands may print the bogus value of the %MEM and RSS, the
reason is that the counter of rss stat is updated in asynchronous manner
and may become negative, when the SPLIT_RSS_COUNTING is enabled in kernel.

As a result, crash will read it from memory and convert from negative to
unsigned long integer, eventually it overflows and gets a big integer. For
example:

  crash> ps 1393
      PID    PPID  CPU       TASK        ST  %MEM      VSZ      RSS  COMM
     1393       1  24  ffff9584bb542100  RU  541298032135.9     4132 18014398509481908  enlinuxpc64
                                             ^^^^^^^^^^^^^^          ^^^^^^^^^^^^^^^^^

This is unexpected, crash needs to correct its value for this case.

Signed-off-by: Lianbo Jiang <lijiang@redhat.com>